requirements.txt
ipykernel
fastapi[standard]
pillow
torch
transformers
ipywidgets
langchain
langchain-community
langchain-ollama